requirements.txt
accelerate
appdirs
loralib
bitsandbytes
black
black[jupyter]
datasets
fire
peft
transformers
sentencepiece
py7zr
scipy
packaging
mosaicml-streaming